Moved core library and split functions out into multiple files
